requirements.txt
requests
psycopg2-binary
pandas
matplotlib
scikit-learn
statsmodels
numpy
python-dateutil
reportlab
pyyaml
